sunday testing changes
fixed drive train logic for not twisting up the wires
added code for shooter control, not working

+++ src/Modules/AnalogChannelVolt.cpp
@@ -48,7 +48,7 @@ float AnalogChannelVolt::GetVoltage()
   return PIDGet();
 }
 
-int AnalogChannelVolt::getturns()
+double AnalogChannelVolt::getturns()
 {
     return motor->GetPosition();
 }
